refactor(github-interactions-api): rename the library to avoid confustion in the nx.json file
