+++ docker-compose.yml
@@ -2,7 +2,6 @@ version: '3.1'
 
 services:
     mongo:
-        
         image: mongo:4.2.6
         ports:
         - '27017:27017'
