+++ libs/api-implementation/orchestration/src/lib/orchestration.controller.ts
@@ -12,7 +12,8 @@ export class VMOrchestrationController {
       await this.orchestrationService.createMachine(createMachineDTO.name);
       return `Machine ${createMachineDTO.name} successfuly created`;
     } catch (error) {
-      return `Error while creating the machine`
+      console.debug(error.stack);
+      return `Error while creating the machine`;
     }
   }
 }
