+++ libs/github-interactions-api-implementation/src/lib/github-interactions-api.module.ts
