+++ libs/github-interactions-api-implementation/tsconfig.json
