+++ libs/github-interactions-api-implementation/tsconfig.lib.json
